updated action on triple quote to add second line
changed the action on triple quote to add a closing quote and
start editing on the newly opened line inside the quotes.  the
only way to jump beyond the quotes is to escape and scroll down.
this is because a <CR> should give a new line, not jump to closing
quotes.