Fix AutoPairsJump for custom pairs
By default AutoPairsJump function is hardcoded for default pairs only,
so it doesn't work for custom pairs and misses '`'
(which is in default AutoPairs).

This commit modifies AutoPairsJump function to properly work with custom
defined pairs by using b:AutoPairs variable values, so that the function
will attempt to search for any closing pair that is configured.